bind_table {textshape}	R Documentation

Column Bind a Table's Values with Its Names

Description

Deprecated, use tidy_table instead.

Usage

bind_table(x, id.name = "id", content.name = "content", ...)

Arguments

x
A table.

id.name
The name to use for the column created from the table names.

content.name
The name to use for the column created from the table values.

...
ignored.

Value

Returns a data.table with the names from the table as an id column.

Examples

## Not run: 
x <- table(sample(LETTERS[1:6], 1000, TRUE))
bind_table(x)

## End(Not run)

[Package textshape version 1.7.5 Index]