hamlet {textshape}	R Documentation

Hamlet (Complete & Split by Sentence)

Description

A dataset containing the complete dialogue of Hamlet with turns of talk split into sentences.

Usage

data(hamlet)

Format

A data frame with 2007 rows and 7 variables

Details

act. The act (akin to repeated measures)

tot. The turn of talk

scene. The scene (nested within an act)

location. Location of the scene

person. Character in the play

died. Logical coded death variable if yes the character dies in the play

dialogue. The spoken dialogue

References

http://www.gutenberg.org

[Package textshape version 1.7.5 Index]