split_run {textshape}	R Documentation

Split Runs

Description

Split runs of consecutive characters.

Usage

split_run(x, ...)

## Default S3 method:
split_run(x, ...)

## S3 method for class 'data.frame'
split_run(x, text.var = TRUE, ...)

Arguments

x
A data.frame or character vector with runs.

text.var
The name of the text variable with runs. If TRUE split_word tries to detect the text column with runs.

...
Ignored.

Value

Returns a list of vectors of runs or an expanded data.table with runs split apart.

Examples

x1 <- c(
     "122333444455555666666",
     NA,
     "abbcccddddeeeeeffffff",
     "sddfg",
     "11112222333"
)

x <- c(rep(x1, 2), ">>???,,,,....::::;[[")

split_run(x)

DATA[["run.col"]] <- x
split_run(DATA, "run.col")

[Package textshape version 1.7.5 Index]