tidy_vector {textshape}	R Documentation

Tidy a Named Atomic Vector: Bind Its Values with Its Names

Description

cbind a named atomic vector's values with its names to form id (from the names) and content columns.

Usage

tidy_vector(x, id.name = "id", content.name = "content", ...)

Arguments

x
A named atomic vector.

id.name
The name to use for the column created from the vector names.

content.name
The name to use for the column created from the vector values.

...
ignored.

Value

Returns a data.table with the names from the vector as an id column.

Examples

x <- setNames(sample(LETTERS[1:6], 1000, TRUE), sample(state.name[1:5], 1000, TRUE))
tidy_vector(x)

[Package textshape version 1.7.5 Index]